requirements.txt
streamlit
torch
torchvision
Pillow
opencv-python
numpy
kagglehub
tqdm
